Add cosmosdb_sql_databases module
A separate module allows the cosmos account to be defined in a different
landing zone to the database. For example you could define a shared cosmos
db account with a private endpoint in level 3 and project-specific sql
databases inside that account in level 4 landing zones. This allows you to
nicely separate shared and project components.

CCOE-191